Mid-fix commit
Still implementing AND operator.

BUG:
If such a program exists in the search queue:
proj(and(o1 : {N1, N2}, o2 : {N2, N3}), ...)
and o1 gets replaced with 'id', giving N1='a', N2='b'.
In the next stage, o2 also gets replaced with id, because N3 can be also ='a'.
But obviously N1 should be different to N3 at all times due to semantics of 'and' operator.